requirements.txt
beautifulsoup4==4.11.1
Django==3.2.6
djangorestframework==3.13.1
fake_useragent==0.1.11
matplotlib==3.5.2
numpy==1.21.2
pandas==1.3.2
pyecharts==1.7.1
PyMySQL==1.0.2
requests==2.28.1
requests_html==0.10.0
scikit_learn==1.1.1
